taskmaster_backend/requirements.txt
Django==4.1.13
djangorestframework==3.12.4
# djongo==1.3.6
django-cors-headers==3.13.0  # Compatible with older Django versions
pytz==2021.3
sqlparse==0.2.4
pymongo==4.10.1
asgiref==3.8.1
djangorestframework-simplejwt==5.3.1